pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vemb"
version = "0.1.0"
description = "httpie for embeddings. Embed anything from the command line."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Yuvraj Angad Singh" }]
keywords = ["embeddings", "gemini", "cli", "rag", "multimodal", "vector"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "google-genai>=1.0.0",
    "click>=8.0.0",
]

[project.scripts]
vemb = "vemb.cli:cli"

[project.urls]
Homepage = "https://github.com/yuvrajangadsingh/vemb"
Repository = "https://github.com/yuvrajangadsingh/vemb"

[tool.pytest.ini_options]
testpaths = ["tests"]
